[fix/#13] move youtubeItem file
