+++ core/src/main/res/drawable/img_bookmark_32.png
