+++ core/src/main/res/drawable/img_bookmark_unselected_32.png
