+++ core/src/main/res/drawable/img_youtube_default_pic.png
